Move rate us / share StoryPad outside to end drawer
