+++ lib/views/community/community_view.dart
@@ -1,14 +1,11 @@
 import 'dart:io';
-
 import 'package:easy_localization/easy_localization.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:storypad/core/constants/app_constants.dart';
-import 'package:storypad/core/extensions/color_scheme_extension.dart';
 import 'package:storypad/core/services/analytics/analytics_service.dart';
-import 'package:storypad/core/services/app_store_opener_service.dart' show AppStoreOpenerService;
 import 'package:storypad/core/services/remote_config/remote_config_service.dart';
 import 'package:storypad/core/services/url_opener_service.dart';
-import 'package:storypad/widgets/bottom_sheets/sp_share_app_bottom_sheet.dart';
+import 'package:storypad/widgets/sp_fade_in.dart';
 import 'package:storypad/widgets/sp_icons.dart';
 import 'package:storypad/widgets/sp_onboarding_wrapper.dart';
 import 'package:storypad/widgets/base_view/view_model_provider.dart';
